rework structure and exports
