+++ src/unitful.jl
@@ -1,7 +1,14 @@
 # previously https://github.com/jw3126/UnitfulRecipes.jl
 # authors: Benoit Pasquier (@briochemc) - Jan Weidner (@jw3126)
 
-using .Unitful: Quantity, unit, ustrip, Unitful, dimension, Units
+module UnitfulRecipes
+
+using ..Unitful: Quantity, unit, ustrip, Unitful, dimension, Units
+using ..RecipesBase
+export @P_str
+
+const AVec = AbstractVector
+const AMat = AbstractMatrix
 
 #==========
 Main recipe
@@ -65,10 +72,7 @@ end
 end
 
 # Recipe for vectors of vectors
-@recipe function f(
-    ::Type{T},
-    x::T,
-) where {T<:AVec{<:AVec{<:Union{Missing,<:Quantity}}}}
+@recipe function f(::Type{T}, x::T) where {T<:AVec{<:AVec{<:Union{Missing,<:Quantity}}}}
     axisletter = plotattributes[:letter]   # x, y, or z
     [fixaxis!(plotattributes, x, axisletter) for x in x]
 end
@@ -188,6 +192,10 @@ end
 Label string containing unit information
 =======================================#
 
+abstract type AbstractProtectedString <: AbstractString end
+struct ProtectedString{S} <: AbstractProtectedString
+    content::S
+end
 struct UnitfulString{S,U} <: AbstractProtectedString
     content::S
     unit::U
@@ -201,6 +209,21 @@ Base.ncodeunits(n::S) = ncodeunits(n.content)
 Base.isvalid(n::S, i::Integer) = isvalid(n.content, i)
 Base.pointer(n::S) = pointer(n.content)
 Base.pointer(n::S, i::Integer) = pointer(n.content, i)
+"""
+    P_str(s)
+
+Creates a string that will be Protected from recipe passes.
+
+Example:
+```julia
+julia> plot(rand(10)*u"m", xlabel=P"This label is protected")
+
+julia> plot(rand(10)*u"m", xlabel=P"This label is not")
+```
+"""
+macro P_str(s)
+    return ProtectedString(s)
+end
 
 #=====================================
 Append unit to labels when appropriate
@@ -236,4 +259,19 @@ format_unit_label(l, u, f::NTuple{2,Char})             = string(l, ' ', f[1], u,
 format_unit_label(l, u, f::NTuple{3,Char})             = string(f[1], l, ' ', f[2], u, f[3])
 format_unit_label(l, u, f::Bool)                       = f ? format_unit_label(l, u, :round) : format_unit_label(l, u, nothing)
 
+const UNIT_FORMATS = Dict(
+    :round => ('(', ')'),
+    :square => ('[', ']'),
+    :curly => ('{', '}'),
+    :angle => ('<', '>'),
+    :slash => '/',
+    :slashround => (" / (", ")"),
+    :slashsquare => (" / [", "]"),
+    :slashcurly => (" / {", "}"),
+    :slashangle => (" / <", ">"),
+    :verbose => " in units of ",
+)
+
 format_unit_label(l, u, f::Symbol) = format_unit_label(l, u, UNIT_FORMATS[f])
+
+end
